tabs: add yet another interface TabViewEngineSession and Obsever
There is EngineSession which be used by SessionManager in AC. In Rocket
current implementation, we had TabView.

To avoid breaking functions exsiting in Rocket, add this abstraction
interface to pretend we don't need TabView in SessionManager.

This layer could be removed once we are ready to get rid of TabView.